Moved unit tests for incomplete JMESPath to the new test file
